requirements.txt
discord.py
Psycopg2
requests
